% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/name_extract.R
\name{name_extract}
\alias{name_extract}
\title{name_extract}
\usage{
name_extract(
  path,
  verify = FALSE,
  language = NULL,
  no_bayes = FALSE,
  odds_details = FALSE,
  data_sources = NULL,
  tokens = NULL
)
}
\arguments{
\item{path}{(character) path to a file}

\item{verify}{(logical) verify names? default: \code{FALSE}}

\item{language}{(logical) text's language. by default it's
automatically detected. default: \code{NULL}}

\item{no_bayes}{(logical) do not run Bayes algorithms. default: \code{FALSE}}

\item{odds_details}{(logical) show details of odds calculation.
default: \code{FALSE}}

\item{data_sources}{(numeric vector) IDs of data sources to display for
matches. default: \code{NULL}. e.g., \code{c(1, 11, 179)}}

\item{tokens}{(logical) xxx. default: \code{NULL}}
}
\value{
list of two data.frames:
\itemize{
\item meta: metadata
\item names: names and their parts, varies based on function parameters
}
}
\description{
extract names using gnfinder
}
\examples{
x <- system.file("examples/BoegerEtal2015.pdf", package="namext")
x
name_extract(x)
name_extract(x, verify = TRUE)
name_extract(x, language = "eng")
name_extract(x, no_bayes = TRUE)
name_extract(x, odds_details = TRUE)
name_extract(x, data_sources = c(4, 12))
name_extract(x, tokens = 5)
}
